requirements.txt
# Core Python dependencies
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1
scikit-learn==1.3.2
matplotlib==3.7.5
seaborn==0.13.2
flask==3.0.3

# Deep Learning & Federated Learning
tensorflow==2.13.0
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
flwr==1.4.0

# Configuration & Utilities
hydra-core==1.3.2
google-auth==2.38.0
requests==2.32.3
